backend/pyproject.toml
[project]
name = "mus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.12",
  "sqlmodel>=0.0.24",
  "uvicorn[standard]>=0.34.2",
  "python-jose[cryptography]>=3.4.0",
  "pytest>=8.3.5",
  "pytest-asyncio>=0.26.0",
  "httpx>=0.28.1",
  "aiosqlite>=0.21.0",
  "greenlet>=3.2.1",
  "pyvips>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
pythonpath = [
  ".", # Add the current directory (backend/) to sys.path
]
